server/src/document_processor.py
"""Document processing module for parsing and chunking documents."""

from typing import List, Dict, Any, Union
from pathlib import Path
import logging
import time
from docx import Document as DocxDocument
import re
from utils.config import MAX_TOKENS, MERGE_PEERS, MIN_CHUNK_SIZE

# Import Docling
from docling.document_converter import DocumentConverter

# Set up logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)


class DocumentProcessor:
    """Handles document conversion and chunking."""
    
    def __init__(self):
        """Initialize the document processor."""
        self.docling_converter = DocumentConverter()
        logger.info("Docling converter initialized successfully")
        
    def _chunk_text(self, text: str, max_tokens: int = MAX_TOKENS, overlap: int = 50) -> List[str]:
        """Simple text chunking by sentences and token count."""
        import re
        
        # Split into sentences
        sentences = re.split(r'[.!?]+', text)
        sentences = [s.strip() for s in sentences if s.strip()]
        
        chunks = []
        current_chunk = ""
        current_tokens = 0
        
        for sentence in sentences:
            sentence_tokens = len(sentence.split())
            
            if current_tokens + sentence_tokens > max_tokens and current_chunk:
                chunks.append(current_chunk.strip())
                # Start new chunk with overlap
                overlap_text = " ".join(current_chunk.split()[-overlap:]) if overlap > 0 else ""
                current_chunk = overlap_text + " " + sentence if overlap_text else sentence
                current_tokens = len(current_chunk.split())
            else:
                current_chunk += " " + sentence
                current_tokens += sentence_tokens
        
        if current_chunk.strip():
            chunks.append(current_chunk.strip())
        
        return chunks
    
    
    def _extract_text_from_docx(self, file_path: str) -> str:
        """Extract text from DOCX file."""
        try:
            doc = DocxDocument(file_path)
            text = ""
            for paragraph in doc.paragraphs:
                text += paragraph.text + "\n"
            return text
        except Exception as e:
            logger.error(f"Error extracting text from DOCX: {e}")
            raise
    
    def _extract_text_from_txt(self, file_path: str) -> str:
        """Extract text from TXT file."""
        try:
            with open(file_path, 'r', encoding='utf-8') as file:
                return file.read()
        except Exception as e:
            logger.error(f"Error extracting text from TXT: {e}")
            raise
    
    def extract_raw_docling_content(self, source: Union[str, Path]) -> str:
        """
        Extract raw content using Docling.
        
        Args:
            source: File path to the document
            
        Returns:
            Raw extracted markdown content
        """
        try:
            logger.info(f"Extracting raw content with Docling: {source}")
            source_path = Path(source)
            
            if not source_path.exists():
                raise FileNotFoundError(f"Document not found: {source}")
            
            # Use Docling for PDFs
            if source_path.suffix.lower() == '.pdf':
                result = self.docling_converter.convert(str(source_path))
                return result.document.export_to_markdown()
            
            # Fallback for other formats
            file_ext = source_path.suffix.lower()
            if file_ext == '.docx':
                return self._extract_text_from_docx(str(source_path))
            elif file_ext in ['.txt', '.md']:
                return self._extract_text_from_txt(str(source_path))
            else:
                raise ValueError(f"Unsupported file format: {file_ext}")
                
        except Exception as e:
            logger.error(f"Error extracting raw content: {e}")
            raise

    def process_document(self, source: Union[str, Path]) -> Dict[str, Any]:
        """
        Process a document from file path.
        
        Args:
            source: File path to the document
            
        Returns:
            Dictionary with chunks and metadata
        """
        try:
            logger.info(f"Processing document: {source}")
            source_path = Path(source)
            
            if not source_path.exists():
                raise FileNotFoundError(f"Document not found: {source}")
            
            # Create md folder if it doesn't exist
            md_folder = Path(__file__).parent.parent / "md"
            md_folder.mkdir(exist_ok=True)
            
            # Extract content using Docling
            logger.info(f"Processing document: {source_path}")
            
            if source_path.suffix.lower() == '.pdf':
                logger.info("Extracting content with Docling...")
                result = self.docling_converter.convert(str(source_path))
                text = result.document.export_to_markdown()
                
                # Save raw Docling extraction to md folder
                md_filename = f"{source_path.stem}_docling_raw.md"
                md_filepath = md_folder / md_filename
                
                logger.info(f"Saving Docling extraction to: {md_filepath}")
                
                with open(md_filepath, 'w', encoding='utf-8') as f:
                    f.write(f"# Raw Docling Extraction\n\n")
                    f.write(f"**Source:** {source_path.name}\n")
                    f.write(f"**Extracted:** {time.strftime('%Y-%m-%d %H:%M:%S')}\n\n")
                    f.write(text)
                
                logger.info(f"✅ Raw Docling extraction saved to: {md_filepath}")
                
            else:
                # Handle other file formats
                file_ext = source_path.suffix.lower()
                if file_ext == '.docx':
                    text = self._extract_text_from_docx(str(source_path))
                elif file_ext in ['.txt', '.md']:
                    text = self._extract_text_from_txt(str(source_path))
                else:
                    raise ValueError(f"Unsupported file format: {file_ext}")
            
            if not text.strip():
                raise ValueError("No text extracted from document")
            
            # Chunk the text
            text_chunks = self._chunk_text(text)
            logger.info(f"Generated {len(text_chunks)} chunks")
            
            # Process chunks into the format we need
            start_time = time.time()
            processed_chunks = []
            for i, chunk_text in enumerate(text_chunks):
                # Filter out very small chunks
                if len(chunk_text.strip()) < MIN_CHUNK_SIZE:
                    continue
                    
                chunk_data = {
                    "text": chunk_text.strip(),
                    "metadata": {
                        "filename": source_path.name,
                        "page_numbers": self._extract_page_numbers_from_text(chunk_text),
                        "title": f"Chunk {i+1}",
                        "source": str(source_path)
                    }
                }
                processed_chunks.append(chunk_data)
            
            logger.info(f"Processed {len(processed_chunks)} valid chunks")
            return {
                'chunks': processed_chunks,
                'full_text': text,  # Store full text for viewing
                'metadata': {
                    'source': source,
                    'total_chunks': len(processed_chunks),
                    'processing_time': time.time() - start_time
                }
            }
            
        except Exception as e:
            logger.error(f"Error processing document {source}: {str(e)}")
            raise
    
    def _extract_page_numbers_from_text(self, text: str) -> List[int]:
        """Extract page numbers from markdown headers or content."""
        import re
        page_numbers = []
        # Look for page references in markdown
        page_matches = re.findall(r'page[\s:]*?(\d+)', text, re.IGNORECASE)
        for match in page_matches:
            page_numbers.append(int(match))
        return sorted(list(set(page_numbers))) if page_numbers else [1]
    
    def process_multiple_documents(self, sources: List[Union[str, Path]]) -> List[Dict[str, Any]]:
        """
        Process multiple documents.
        
        Args:
            sources: List of file paths or URLs
            
        Returns:
            Combined list of processed chunks from all documents
        """
        all_chunks = []
        
        for source in sources:
            try:
                chunks = self.process_document(source)
                all_chunks.extend(chunks)
            except Exception as e:
                logger.error(f"Failed to process {source}: {str(e)}")
                continue
        
        return all_chunks
