server/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "docparses-server"
version = "0.1.0"
description = "Document Parser and RAG Pipeline Server"
authors = [{name = "Your Name", email = "your.email@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "sentence-transformers==3.0.1",
    "chromadb==0.4.22",
    "groq==0.9.0",
    "langchain-core==0.3.63",
    "langchain-google-genai==2.0.5",
    "fastapi==0.115.0",
    "uvicorn==0.32.0",
    "python-multipart==0.0.12",
    "python-dotenv==1.0.1",
    "pandas==2.2.2",
    "pydantic==2.8.2",
    "torch==2.4.0",
    "transformers==4.44.0",
    "numpy==1.26.4",
    "python-docx==1.1.2",
    "google-generativeai==0.8.3",
]

[project.urls]
Homepage = "https://github.com/yourusername/docparses-server"
Repository = "https://github.com/yourusername/docparses-server"

[tool.uvicorn]
host = "0.0.0.0"
port = 8000
reload = true
